Refactor move code from manager to model module
